+++ app/src/main/java/com/eyeson/android/ui/meeting/MeetingViewModel.kt
@@ -317,6 +317,7 @@ class MeetingViewModel @Inject constructor(
 
         override fun onCameraError(error: String) {
             addEvent("onCameraError: error $error", true)
+            cameraDisconnected.value = true
         }
 
         override fun onCameraSwitchDone(isFrontCamera: Boolean) {